+++ FirebaseAI/Tests/Unit/GenerativeModelSessionTests.swift
@@ -25,7 +25,7 @@
     let testPrompt = "What sorts of questions can I ask you?"
     let testModelName = "test-model"
     let testModelResourceName = "projects/test-project-id/models/test-model"
-    let apiConfig = FirebaseAI.defaultVertexAIAPIConfig
+    let apiConfig = FirebaseAI.defaultEnterpriseAPIConfig
 
     let googleAISubdirectory = "mock-responses/googleai"